% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-attributes.R
\name{gpkg_write_attributes}
\alias{gpkg_write_attributes}
\alias{gpkg_remove_attributes}
\title{Write or Remove Attribute Table in a GeoPackage}
\usage{
gpkg_write_attributes(
  x,
  table,
  table_name,
  description = "",
  template = NULL,
  overwrite = FALSE,
  append = FALSE
)

gpkg_remove_attributes(x, table_name)
}
\arguments{
\item{x}{a geopackage object}

\item{table}{a data.frame}

\item{table_name}{character. The name for \code{table} in \code{x}}

\item{description}{Optional description. Default \code{""}}

\item{template}{a list (containing elements \code{"ext"} and \code{"crs"}, or a terra object. These are defining xmin/ymin/xmax/ymax and CRS of the attribute table.}

\item{overwrite}{overwrite? Default \code{FALSE}}

\item{append}{append? Default \code{FALSE}}
}
\description{
\code{gpkg_write_attributes()}: Specify a target geopackage and name for new table. For adding attributes, specify the new data as data.frame. The table name will be registered in the \code{gpkg_contents} table. Optionally include a description and/or a template terra object that can be used to define the extent of the attribute data.

\code{gpkg_remove_attributes()}: Remove an attribute table and corresponding \code{gpkg_contents} record
}
