% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-connection.R
\name{gpkg_connect}
\alias{gpkg_connect}
\alias{gpkg_connect.geopackage}
\alias{gpkg_connect.character}
\alias{gpkg_is_connected}
\alias{gpkg_is_connected.geopackage}
\alias{gpkg_disconnect}
\alias{gpkg_disconnect.geopackage}
\alias{gpkg_disconnect.SQLiteConnection}
\title{Create SQLite Connection to GeoPackage}
\usage{
gpkg_connect(x)

\method{gpkg_connect}{geopackage}(x)

\method{gpkg_connect}{character}(x)

gpkg_is_connected(x)

\method{gpkg_is_connected}{geopackage}(x)

gpkg_disconnect(x)

\method{gpkg_disconnect}{geopackage}(x)

\method{gpkg_disconnect}{SQLiteConnection}(x)
}
\arguments{
\item{x}{A \code{geopackage} or \code{SQLiteConnection} object}
}
\value{
A DBIConnection (SQLiteConnection) object. \code{NULL} on error.

Logical (invisible). \code{FALSE} if connection cannot be closed.
}
\description{
Create SQLite Connection to GeoPackage

Create SQLite Connection to GeoPackage
}
