% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-package.R
\docType{package}
\name{gpkg-package}
\alias{gpkg}
\alias{gpkg-package}
\title{gpkg: Proof of Concept 'GeoPackage' to Arrow Converter}
\description{
Provides a proof-of-concept converter that takes a gpkg query and returns the data using the 'Arrow' C data interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/gpkg}
  \item Report bugs at \url{https://github.com/paleolimbot/gpkg/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
