% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-util.R
\name{gpkg_2d_gridded_coverage_ancillary}
\alias{gpkg_2d_gridded_coverage_ancillary}
\title{Get \code{gpkg_2d_gridded_coverage_ancillary} Table}
\usage{
gpkg_2d_gridded_coverage_ancillary(x)
}
\arguments{
\item{x}{A a \code{geopackage} object, path to a GeoPackage or an \code{SQLiteConnection}}
}
\value{
a data.frame containing columns \code{id}, \code{tile_matrix_set_name}, \code{datatype}, \code{scale}, \code{offset}, \code{precision}, \code{data_null}, \code{grid_cell_encoding}, \code{uom}, \code{field_name}, \code{quantity_definition}
}
\description{
Get \code{gpkg_2d_gridded_coverage_ancillary} Table
}
