% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-contents.R
\name{gpkg_add_contents}
\alias{gpkg_add_contents}
\title{Add a record to \code{gpkg_contents}}
\usage{
gpkg_add_contents(x, table_name, description = "", template = NULL)
}
\arguments{
\item{x}{A \emph{geopackage}}

\item{table_name}{Name of table to add record for in gpkg_contents}

\item{description}{Default \code{""}}

\item{template}{Default \code{NULL} uses global EPSG:4326 with bounds -180,-90:180,90}
}
\value{
A \emph{geopackage}
}
\description{
Add a record to \code{gpkg_contents}
}
