% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-extensions.R
\name{gpkg_add_metadata_extension}
\alias{gpkg_add_metadata_extension}
\title{Add 'Metadata' extension}
\usage{
gpkg_add_metadata_extension(x)
}
\arguments{
\item{x}{a \code{geopackage}}
}
\value{
\code{0} (invisible). Called for side effects.
}
\description{
Adds the "Metadata" extension tables.
}
