% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-datasets.R
\docType{data}
\name{gpkg_creation_options}
\alias{gpkg_creation_options}
\title{GeoPackage Creation Options}
\format{
An object of class \code{data.frame} with 21 rows and 4 columns.
}
\source{
\href{https://gdal.org/drivers/raster/gpkg.html#creation-options}{GDAL - GPKG – GeoPackage raster}
}
\usage{
gpkg_creation_options
}
\description{
GeoPackage Creation Options
}
\keyword{datasets}
