% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{gpkg_example}
\alias{gpkg_example}
\title{GeoPackage Example Files}
\usage{
gpkg_example(name)
}
\arguments{
\item{name}{The example name.}
}
\value{
A filename or an error if the example does not exist
}
\description{
Current example files are 'nc', geometrycollection_m', 'geometrycollection_z',
'geometrycollection_zm', 'geometrycollection', 'linestring_m',
'linestring_z', 'linestring_zm', 'linestring', 'multilinestring_m',
'multilinestring_z', 'multilinestring_zm', 'multilinestring', 'multipoint_m',
'multipoint_z', 'multipoint_zm', 'multipoint', 'multipolygon_m',
'multipolygon_z', 'multipolygon_zm', 'multipolygon', 'point_m',
'point_z', 'point_zm', 'point', 'polygon_m', 'polygon_z', 'polygon_zm',
and 'polygon'.
}
\examples{
gpkg_example("nc")

}
