% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-execute.R
\name{gpkg_execute}
\alias{gpkg_execute}
\title{Execute an SQL statement in a GeoPackage}
\usage{
gpkg_execute(x, statement)
}
\arguments{
\item{x}{a \emph{geopackage} object}

\item{statement}{a SQLite statement}
}
\value{
result of \code{RSQLite::dbExecute()}
}
\description{
Execute an SQL statement in a GeoPackage
}
