% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-extensions.R
\name{gpkg_extensions}
\alias{gpkg_extensions}
\title{Get Geopackage Extensions}
\usage{
gpkg_extensions(x)
}
\arguments{
\item{x}{A \code{geopackage}}
}
\value{
a data.frame
}
\description{
Get Geopackage Extensions
}
