% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-contents.R
\name{gpkg_list_contents}
\alias{gpkg_list_contents}
\title{List Tables Registered in a GeoPackage \code{gpkg_contents}}
\usage{
gpkg_list_contents(x, ogr = FALSE)
}
\arguments{
\item{x}{A \emph{geopackage} object, path to a GeoPackage or an \emph{SQLiteConnection}}

\item{ogr}{Intersect \code{gpkg_contents} table name result with OGR tables that are listed in \code{gpkg_ogr_contents}? Default: \code{FALSE}}
}
\value{
character. Vector of grid, feature and attribute table names registered in GeoPackage.
}
\description{
Get a vector of grid, feature and attribute table names registered in GeoPackage.
}
\seealso{
\code{\link[=gpkg_contents]{gpkg_contents()}} \code{\link[=gpkg_list_tables]{gpkg_list_tables()}}
}
