% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-util.R
\name{gpkg_list_tables}
\alias{gpkg_list_tables}
\title{List Tables in a GeoPackage}
\usage{
gpkg_list_tables(x)
}
\arguments{
\item{x}{A a \code{geopackage}, path to a GeoPackage or an \code{SQLiteConnection}}
}
\value{
a character vector with names of all tables and views in the database
}
\description{
List Tables in a GeoPackage
}
