% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg.R
\name{gpkg_open}
\alias{gpkg_open}
\alias{gpkg_close}
\alias{gpkg_exec}
\alias{gpkg_query}
\alias{gpkg_query_narrow}
\alias{gpkg_query_table}
\alias{gpkg_open_test}
\alias{gpkg_list_tables}
\title{SQLite3 Interface}
\usage{
gpkg_open(file = ":memory:")

gpkg_close(con)

gpkg_exec(con, sql)

gpkg_query(con, sql)

gpkg_query_narrow(con, sql)

gpkg_query_table(con, sql)

gpkg_open_test()

gpkg_list_tables(con)
}
\arguments{
\item{file}{A filename or :memory: for an in-memory database.}

\item{con}{A connection opened using \code{\link[=gpkg_open]{gpkg_open()}}}

\item{sql}{An SQL statement}
}
\description{
SQLite3 Interface
}
