% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{gpkg_read}
\alias{gpkg_read}
\title{Read a GeoPackage}
\usage{
gpkg_read(dsn)
}
\arguments{
\item{dsn}{Path to GeoPackage}
}
\value{
List containing tables and references to grids
}
\description{
Read a GeoPackage
}
\keyword{io}
