% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-util.R
\name{gpkg_source}
\alias{gpkg_source}
\alias{gpkg_source.geopackage}
\title{Get Source File of a \code{geopackage} Object}
\usage{
gpkg_source(x)

\method{gpkg_source}{geopackage}(x)
}
\arguments{
\item{x}{a \code{geopackage} object}
}
\value{
character path
}
\description{
Get Source File of a \code{geopackage} Object
}
