% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-datasets.R
\docType{data}
\name{gpkg_sqlite_tables}
\alias{gpkg_sqlite_tables}
\title{GeoPackage SQLite Tables}
\format{
a \code{data.frame} with 1 column (\code{"table_name"}) and 10 rows
}
\source{
\href{https://gdal.org/drivers/raster/gpkg.html#creation-options}{GDAL - GPKG – GeoPackage raster}
}
\usage{
gpkg_sqlite_tables
}
\description{
GeoPackage SQLite Tables
}
\keyword{datasets}
