% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-sqlite.R
\name{gpkg_table}
\alias{gpkg_table}
\title{general function to get table by table name}
\usage{
gpkg_table(x, table_name, query_string = FALSE)
}
\arguments{
\item{x}{A a \code{geopackage} object, path to a GeoPackage or an \code{SQLiteConnection}}

\item{table_name}{character. table name}

\item{query_string}{logical. Return SQLite query rather than executing it? Default: \code{FALSE}}
}
\value{
A data.frame
}
\description{
general function to get table by table name
}
