% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-util.R
\name{gpkg_tile_set_data_null}
\alias{gpkg_tile_set_data_null}
\title{Set \code{data_null} Metadata for a GeoPackage Tile Dataset}
\usage{
gpkg_tile_set_data_null(x, name, value, query_string = FALSE)
}
\arguments{
\item{x}{A \emph{geopackage} object, path to a GeoPackage or an \emph{SQLiteConnection}}

\item{name}{character. Tile matrix set name(s) (\code{tile_matrix_set_name})}

\item{value}{numeric. Value to use as "NoData" (\code{data_null} value)}

\item{query_string}{logical. Return SQLite query rather than executing it? Default: \code{FALSE}}
}
\value{
logical. \code{TRUE} if number of \code{data_null} records updated is greater than \code{0}.
}
\description{
Set \code{data_null} Metadata for a GeoPackage Tile Dataset
}
