% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-contents.R
\name{gpkg_update_contents}
\alias{gpkg_update_contents}
\title{Build \code{gpkg_contents} table based on existing tables}
\usage{
gpkg_update_contents(x)
}
\arguments{
\item{x}{a \emph{geopackage} object}
}
\value{
logical. \code{TRUE} on successful update (all non-standard tables added to gpkg_contents table)
}
\description{
Build \code{gpkg_contents} table based on existing tables
}
