% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-attributes.R
\name{gpkg_write_attributes}
\alias{gpkg_write_attributes}
\title{Write an attribute table to GeoPackage}
\usage{
gpkg_write_attributes(
  x,
  table,
  table_name,
  description = "",
  template = NULL,
  overwrite = FALSE,
  append = FALSE
)
}
\arguments{
\item{x}{a geopackage object}

\item{table}{a data.frame}

\item{table_name}{character. The name for \code{table} in \code{x}}

\item{description}{Optional description. Default \code{""}}

\item{template}{a list (containing elements \code{"ext"} and \code{"crs"}, or a terra object. These are defining xmin/ymin/xmax/ymax and CRS of the attribute table.}

\item{overwrite}{overwrite? Default \code{FALSE}}

\item{append}{append? Default \code{FALSE}}
}
\description{
Specify a target geopackage, data.frame of input data, and name for table.
}
