% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-2d-gridded-ancillary.R
\name{gpkg_create_2d_gridded_coverage_ancillary}
\alias{gpkg_create_2d_gridded_coverage_ancillary}
\alias{gpkg_create_2d_gridded_tile_ancillary}
\title{Create 2D Gridded Coverage and Tile Ancillary Tables}
\usage{
gpkg_create_2d_gridded_coverage_ancillary(x)

gpkg_create_2d_gridded_tile_ancillary(x)
}
\arguments{
\item{x}{A \emph{geopackage} object}
}
\value{
\code{gpkg_create_2d_gridded_coverage_ancillary()}: \emph{integer}. Result of running sequential \code{gpkg_execute()} statements.

\code{gpkg_create_2d_gridded_tile_ancillary()}: \emph{integer}. Result of running sequential \code{gpkg_execute()} statements.
}
\description{
Create 2D Gridded Coverage and Tile Ancillary Tables
}
\references{
http://www.opengis.net/doc/IS/geopackage-gr/1.0
}
