% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_gpkg_sf}
\alias{read_gpkg_sf}
\alias{read_gpkg_table}
\alias{gpkg_contents}
\title{Read a GeoPackage}
\usage{
read_gpkg_sf(file, layer = NULL)

read_gpkg_table(file, layer = NULL)

gpkg_contents(file)
}
\arguments{
\item{file}{A path to a .gpkg file}

\item{layer}{A name of a layer or NULL for the first layer}
}
\value{
\itemize{
\item \code{read_gpkg_sf()} returns an \link[sf:sf]{sf object}
\item \code{read_gpkg_table()} returns an \link[arrow:Table]{arrow::Table}
}
}
\description{
Read a GeoPackage
}
\examples{
\dontshow{if (requireNamespace("arrow", quietly = TRUE) && requireNamespace("sf", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_gpkg_sf(gpkg_example("nc"))
\dontshow{\}) # examplesIf}
}
