% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrowsqlite3.R
\name{sqlite3_open}
\alias{sqlite3_open}
\alias{sqlite3_close}
\alias{sqlite3_exec}
\alias{sqlite3_query}
\alias{sqlite3_query_narrow}
\alias{sqlite3_query_table}
\alias{sqlite3_open_test}
\alias{sqlite3_list_tables}
\title{SQLite3 Interface}
\usage{
sqlite3_open(file = ":memory:")

sqlite3_close(con)

sqlite3_exec(con, sql)

sqlite3_query(con, sql)

sqlite3_query_narrow(con, sql)

sqlite3_query_table(con, sql)

sqlite3_open_test()

sqlite3_list_tables(con)
}
\arguments{
\item{file}{A filename or :memory: for an in-memory database.}

\item{con}{A connection opened using \code{\link[=sqlite3_open]{sqlite3_open()}}}

\item{sql}{An SQL statement}
}
\description{
SQLite3 Interface
}
