% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-class.R
\name{geopackage}
\alias{geopackage}
\alias{geopackage.list}
\alias{geopackage.missing}
\alias{geopackage.SQLiteConnection}
\alias{geopackage.geopackage}
\alias{geopackage.character}
\title{\code{geopackage} Constructors}
\usage{
geopackage(x, ...)

\method{geopackage}{list}(x, dsn = NULL, connect = FALSE, ...)

\method{geopackage}{missing}(x, connect = FALSE, pattern = "Rgpkg", tmpdir = tempdir(), ...)

\method{geopackage}{SQLiteConnection}(x, connect = FALSE, ...)

\method{geopackage}{geopackage}(x, ...)

\method{geopackage}{character}(x, connect = FALSE, ...)
}
\arguments{
\item{x}{list of SpatVectorProxy, SpatRaster, data.frame; or a character containing path to a GeoPackage file; or an SQLiteConnection to a GeoPackage. If missing, a temporary file with .gpkg extension is created in \code{tempdir}.}

\item{...}{Additional arguments [not currently used]}

\item{dsn}{Path to GeoPackage File (may not exist)}

\item{connect}{Connect to database and store connection in result? Default: \code{FALSE}}

\item{pattern}{used only when \code{x} is missing (creating temporary file GeoPackage), passed to \code{tempfile()}; default \code{"Rgpkg"}}

\item{tmpdir}{used only when \code{x} is missing (creating temporary file GeoPackage), passed to \code{tempfile()}; default \code{tempdir()}}
}
\value{
A \code{geopackage} Object
}
\description{
\code{geopackage} Constructors
}
