% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-connection.R
\name{gpkg_connect}
\alias{gpkg_connect}
\alias{gpkg_connect.geopackage}
\alias{gpkg_connect.character}
\alias{gpkg_is_connected}
\alias{gpkg_is_connected.geopackage}
\alias{gpkg_disconnect}
\alias{gpkg_disconnect.geopackage}
\alias{gpkg_disconnect.SQLiteConnection}
\title{Create SQLite Connection to GeoPackage}
\usage{
gpkg_connect(x)

\method{gpkg_connect}{geopackage}(x)

\method{gpkg_connect}{character}(x)

gpkg_is_connected(x)

\method{gpkg_is_connected}{geopackage}(x)

gpkg_disconnect(x)

\method{gpkg_disconnect}{geopackage}(x)

\method{gpkg_disconnect}{SQLiteConnection}(x)
}
\arguments{
\item{x}{A \emph{geopackage} or \emph{SQLiteConnection} object}
}
\value{
A DBIConnection (SQLiteConnection) object. \code{NULL} on error.

If \code{x} is \emph{geopackage}, the disconnected object is returned. If x is a \emph{SQLiteConnection}, logical (\code{TRUE} if successfully disconnected).
}
\description{
Method for creating and connecting \code{SQLiteConnection} object stored within \code{geopackage} object.
}
\details{
The S3 method for \code{geopackage} objects uses in-place modification to update the parent object by name. That is, if you call \code{gpkg_connect()} on an object \code{g} then as a side-effect \code{g} is updated in the calling environment. This behavior is considered by many to be non-idiomatic for R, but it is useful to provide a simple way to connect an existing object without having to retain references to pointers to connection objects. To avoid replacement of object values in the parent frame, you can use the \code{character} method. That is, \code{g <- gpkg_connect(g$dsn)} is equivalent to \code{gpkg_connect(g)} when \code{g} is a \code{geopackage}.
}
