% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-package.R
\docType{package}
\name{gpkg-package}
\alias{gpkg}
\alias{gpkg-package}
\title{gpkg: Utilities for OGC GeoPackages}
\description{
High-level wrapper functions to build GeoPackages containing a variety of different data. Read/write of spatial vector and gridded data is performed using standard GDAL utilities. Additional functions are provided to manipulate attributes and tabular data.
}
\keyword{internal}
