% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-package.R
\docType{package}
\name{gpkg-package}
\alias{gpkg}
\alias{gpkg-package}
\title{gpkg: Utilities for the Open Geospatial Consortium 'GeoPackage' Format}
\description{
Build Open Geospatial Consortium 'GeoPackage' files (\url{https://www.geopackage.org/}). 'GDAL' utilities for reading and writing spatial data are provided by the 'terra' package. Additional 'GeoPackage' and 'SQLite' features for attributes and tabular data are implemented with the 'RSQLite' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://humus.rocks/gpkg/}
  \item \url{https://github.com/brownag/gpkg}
  \item Report bugs at \url{https://github.com/brownag/gpkg/issues}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{brown.andrewg@gmail.com}

}
\keyword{internal}
