% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-extensions.R
\name{gpkg_add_metadata_extension}
\alias{gpkg_add_metadata_extension}
\title{Add Metadata extension}
\usage{
gpkg_add_metadata_extension(x)
}
\arguments{
\item{x}{a \code{geopackage}}
}
\description{
Add Metadata extension
}
