% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-extensions.R
\name{gpkg_add_relatedtables_extension}
\alias{gpkg_add_relatedtables_extension}
\title{Add Related Tables extension}
\usage{
gpkg_add_relatedtables_extension(x)
}
\arguments{
\item{x}{a \code{geopackage}}
}
\description{
Add Related Tables extension
}
