% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-contents.R
\name{gpkg_contents}
\alias{gpkg_contents}
\alias{gpkg_ogr_contents}
\title{Get \code{gpkg_contents} or \code{gpkg_ogr_contents} Table}
\usage{
gpkg_contents(x, create = FALSE)

gpkg_ogr_contents(x)
}
\arguments{
\item{x}{A \emph{geopackage} object, path to a GeoPackage or an \emph{SQLiteConnection}}

\item{create}{Create table \code{gpkg_contents} if does not exist? Default: ``}
}
\value{
\code{gpkg_contents()}: a \emph{data.frame} containing columns \code{table_name}, \code{data_type}, \code{identifier}, \code{description}, \code{last_change}, \code{min_x}, \code{min_y}, \code{max_x}, \code{max_y}, \code{srs_id}

\code{gpkg_ogr_contents()}: a \emph{data.frame} containing columns \code{table_name} and \code{feature_count}.
}
\description{
These functions provide convenient access to the contents of the standard GeoPackage tables of the same name.
}
