% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{gpkg_example}
\alias{gpkg_example}
\title{GeoPackage Example Files}
\usage{
gpkg_example(name)
}
\arguments{
\item{name}{The example name.}
}
\value{
A filename or an error if the example does not exist
}
\description{
GeoPackage Example Files
}
\examples{
gpkg_example("point")

}
