% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-execute.R
\name{gpkg_execute}
\alias{gpkg_execute}
\title{Execute an SQL statement in a GeoPackage}
\usage{
gpkg_execute(x, statement, ..., silent = FALSE)
}
\arguments{
\item{x}{a \emph{geopackage} object}

\item{statement}{a SQLite statement}

\item{...}{Additional arguments to \code{RSQLite::dbExecute()}}

\item{silent}{Used to suppress error messages, passed to \code{try()}. Default: \code{FALSE}.}
}
\value{
Invisible result of \code{RSQLite::dbExecute()}; or \code{try-error} on error.
}
\description{
Execute an SQL statement in a GeoPackage
}
