% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-query.R
\name{gpkg_query}
\alias{gpkg_query}
\title{Query a GeoPackage for tabular result}
\usage{
gpkg_query(x, query)
}
\arguments{
\item{x}{a \emph{geopackage} object}

\item{query}{a SQLite query}
}
\value{
a \emph{data.frame} result of \code{RSQLite::dbGetQuery()}
}
\description{
Query a GeoPackage for tabular result
}
