% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-io.R
\name{gpkg_read}
\alias{gpkg_read}
\title{Read data from a GeoPackage}
\usage{
gpkg_read(x, connect = FALSE, quiet = TRUE)
}
\arguments{
\item{x}{Path to GeoPackage}

\item{connect}{Connect to database and store connection in result? Default: \code{FALSE}}

\item{quiet}{Hide printing of gdalinfo description to stdout. Default: \code{TRUE}}
}
\value{
A \code{geopackage} object (list containing tables, grids and vector data)
}
\description{
Experimental: This function is being evaluated for its scope compared to other more general functions that perform similar operations (i.e. \code{gpkg_tables()}).
}
\keyword{io}
