% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-util.R
\name{gpkg_source}
\alias{gpkg_source}
\alias{gpkg_source.geopackage}
\title{Get Source File of a \emph{geopackage} object}
\usage{
gpkg_source(x)

\method{gpkg_source}{geopackage}(x)
}
\arguments{
\item{x}{A \emph{geopackage} object}
}
\value{
\emph{character} file path
}
\description{
Get Source File of a \emph{geopackage} object
}
