% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-datasets.R
\docType{data}
\name{gpkg_sqlite_tables}
\alias{gpkg_sqlite_tables}
\title{GeoPackage Dataset}
\format{
a \code{data.frame} with 1 column (\code{"table_name"}) and 10 rows
}
\source{
\href{https://gdal.org/drivers/raster/gpkg.html}{GDAL - GPKG – GeoPackage raster}, \href{https://gdal.org/drivers/vector/gpkg.html}{GDAL - GPKG – GeoPackage vector}
}
\usage{
gpkg_sqlite_tables
}
\description{
GeoPackage Dataset

GeoPackage SQLite Tables
}
\keyword{datasets}
