% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-util.R
\name{gpkg_tables}
\alias{gpkg_tables}
\alias{gpkg_tables.geopackage}
\title{Get Tables from a \code{geopackage} Object}
\usage{
gpkg_tables(x)

\method{gpkg_tables}{geopackage}(x)
}
\arguments{
\item{x}{a \code{geopackage} object}
}
\value{
a list of SpatVectorProxy, SpatRaster, data.frame (lazy tbl?)
}
\description{
Get Tables from a \code{geopackage} Object
}
