% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-util.R
\name{gpkg_tables}
\alias{gpkg_tables}
\alias{gpkg_tables.geopackage}
\title{Get Tables from a \code{geopackage} Object}
\usage{
gpkg_tables(x, collect = FALSE, pragma = FALSE)

\method{gpkg_tables}{geopackage}(x, collect = FALSE, pragma = FALSE)
}
\arguments{
\item{x}{a \code{geopackage} object}

\item{collect}{Default: \code{FALSE}. Should tables be materialized as 'data.frame' objects in memory? (i.e. not "lazy") Default: \code{FALSE}; if \code{TRUE} 'dbplyr' is not required. Always \code{TRUE} for \code{pragma=TRUE} (pragma information are always "collected").}

\item{pragma}{Default: \code{FALSE}. Use \code{gpkg_table_pragma()} instead of \code{gpkg_table()}? The former does not require 'dbplyr'.}
}
\value{
a list of SpatVectorProxy, SpatRaster, data.frame (lazy tbl?)
}
\description{
Get Tables from a \code{geopackage} Object
}
