% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-sqlite.R
\name{gpkg_update_table}
\alias{gpkg_update_table}
\title{Update Table by Name}
\usage{
gpkg_update_table(
  x,
  table_name,
  updatecol,
  updatevalue,
  wherecol,
  wherevector,
  query_string = FALSE
)
}
\arguments{
\item{x}{A a \code{geopackage} object, path to a GeoPackage or an \code{SQLiteConnection}}

\item{table_name}{character. table name}

\item{updatecol}{character. column to update}

\item{updatevalue}{scalar value to set}

\item{wherecol}{character. column to constrain update}

\item{wherevector}{vector of values where update should be made}

\item{query_string}{logical. Return SQLite query rather than executing it? Default: \code{FALSE}}
}
\description{
For a given table, set column X to scalar A where column Y is in vector B
}
