% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-validate.R
\name{gpkg_validate}
\alias{gpkg_validate}
\title{Validate a GeoPackage}
\usage{
gpkg_validate(x, diagnostics = FALSE)
}
\arguments{
\item{x}{Path to GeoPackages}

\item{diagnostics}{Return a list containing diagnostics (missing table names, invalid values, other errors)}
}
\value{
\code{TRUE} if valid. \code{FALSE} if one or more problems are found. For full diagnostics run with \code{diagnostics = TRUE} to return a list containing results for each input GeoPackage.
}
\description{
Checks for presence of required tables, valid values and other constraints.
}
