% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{gpkg_write}
\alias{gpkg_write}
\title{Write a GeoPackage}
\usage{
gpkg_write(dsn, destfile)
}
\arguments{
\item{dsn}{Character. Source file path(s)}

\item{destfile}{Path to output GeoPackage file}
}
\value{
Logical. `TRUE` on success.
}
\description{
Write a GeoPackage
}
\keyword{io}
