% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg-lazy.frame.R
\name{lazy.frame}
\alias{lazy.frame}
\alias{lazy.frame.character}
\alias{lazy.frame.geopackage}
\title{lazy data.frame
get information on a table in a GeoPackage (without returning the whole table)}
\usage{
lazy.frame(x, ...)

\method{lazy.frame}{character}(x, ...)

\method{lazy.frame}{geopackage}(x, table = NULL, ...)
}
\arguments{
\item{x}{a geopackage object or character path to GeoPcakge}

\item{...}{additional arguments not used}

\item{table}{one or more table names}
}
\description{
lazy data.frame
get information on a table in a GeoPackage (without returning the whole table)
}
